requirements.txt
Flask==3.0.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
eventlet==0.36.1
werkzeug==3.0.3
PyMySQL==1.1.0
Flask-Login==0.6.3
